rwlock: add prioritize_write flag to prevent write starvation
RWLock's constructor takes an additional 'prioritize_write' flag, which
defaults to false.  when enabled, this adds the rwlock attribute
PTHREAD_RWLOCK_PREFER_WRITER_NONRECURSIVE_NP, which will block calls to
pthread_rwlock_rdlock() if there are any writers waiting.  this avoids
starvation of pthread_rwlock_wrlock() in read-mostly workloads

also made the destructor non-virtual; this was adding a virtual function
table to the class, even though there are no other virtual functions.
added the 'final' keyword to document that it's not meant to be derived
from

Signed-off-by: Casey Bodley <cbodley@redhat.com>